refactor(policies): extract AST types into shared.policies.ast
Move Expr/Policy/PolicyDocument and the AST validator out of
src/models/contracts/policies.py and into shared/policies/ast.py.
The contracts file becomes a shim that re-exports KNOWN_USER_FIELDS,
Expr, and defines a tables-narrowed Policy + TablePolicies preserving
the table action vocab for existing callers.
